# Add basic benchmarks for outputs, unspent outputs, and transactions