--- a/persist/sqlite/v2consensus_refactored_test.go
+++ b/persist/sqlite/v2consensus_refactored_test.go
@@ -1,4 +1,4 @@
-package sqlite_test
+package sqlite
 
 import (
 	"errors"