Latest benchmarks:
```
BenchmarkTransactions/10_transactions
BenchmarkTransactions/10_transactions-4                      940           1219615 ns/op
BenchmarkTransactions/100_transactions
BenchmarkTransactions/100_transactions-4                     142           8281356 ns/op
BenchmarkTransactions/1000_transactions
BenchmarkTransactions/1000_transactions-4                     14          78907892 ns/op

BenchmarkSiacoinOutputs/10_unspent_outputs
BenchmarkSiacoinOutputs/10_unspent_outputs-4                6830            230802 ns/op
BenchmarkSiacoinOutputs/100_unspent_outputs
BenchmarkSiacoinOutputs/100_unspent_outputs-4               1513            862293 ns/op
BenchmarkSiacoinOutputs/1000_unspent_outputs
BenchmarkSiacoinOutputs/1000_unspent_outputs-4               164           7006526 ns/op

BenchmarkSiacoinOutputs/10_siacoin_elements
BenchmarkSiacoinOutputs/10_siacoin_elements-4               4574            272862 ns/op
BenchmarkSiacoinOutputs/100_siacoin_elements
BenchmarkSiacoinOutputs/100_siacoin_elements-4               968           1285014 ns/op
BenchmarkSiacoinOutputs/1000_siacoin_elements
BenchmarkSiacoinOutputs/1000_siacoin_elements-4               96          12238006 ns/op
```
https://github.com/SiaFoundation/explored/issues/201