@n8maninger 2025-05-30
> I suspect the bulk of the difference between the unspent outputs and siacoin elements queries is that we retrieve the merkle proof in `SiacoinElements` which can require multiple queries. Not sure if we want to do anything about it, we could try going back to a separate on disk representation of the proofs instead of using sqlite.

The problem isn't the proof storage, it's that the retrieval is inefficient. batching + prepared statement + caching provides a significant speed up:

```
diff --git a/persist/sqlite/addresses.go b/persist/sqlite/addresses.go
index 8f2e3e7..0cee8c9 100644
--- a/persist/sqlite/addresses.go
+++ b/persist/sqlite/addresses.go
@@ -4,6 +4,7 @@ import (
 	"database/sql"
 	"errors"
 	"fmt"
+	"math/bits"
 
 	"go.sia.tech/core/types"
 	"go.sia.tech/explored/explorer"
@@ -154,6 +155,50 @@ func (s *Store) UnspentSiafundOutputs(address types.Address, offset, limit uint6
 	return
 }
 
+func fillElementProofs(tx *txn, indices []uint64) (proofs [][]types.Hash256, _ error) {
+	if len(indices) == 0 {
+		return nil, nil
+	}
+
+	var numLeaves uint64
+	if err := tx.QueryRow(`SELECT num_leaves FROM network_metrics ORDER BY height DESC LIMIT 1`).Scan(decode(&numLeaves)); err != nil {
+		return nil, fmt.Errorf("failed to query state tree leaves: %w", err)
+	}
+
+	stmt, err := tx.Prepare(`SELECT value FROM state_tree WHERE row=? AND column=?`)
+	if err != nil {
+		return nil, fmt.Errorf("failed to prepare statement: %w", err)
+	}
+	defer stmt.Close()
+
+	data := make(map[uint64]map[uint64]types.Hash256)
+	for _, leafIndex := range indices {
+		proof := make([]types.Hash256, bits.Len64(leafIndex^numLeaves)-1)
+		for j := range proof {
+			row, col := uint64(j), (leafIndex>>j)^1
+
+			// check if the hash is already in the cache
+			if h, ok := data[row][col]; ok {
+				proof[j] = h
+				continue
+			}
+
+			// query the hash from the database
+			if err := stmt.QueryRow(row, col).Scan(decode(&proof[j])); err != nil {
+				return nil, fmt.Errorf("failed to query state element (%d,%d): %w", row, col, err)
+			}
+
+			// cache the hash
+			if _, ok := data[row]; !ok {
+				data[row] = make(map[uint64]types.Hash256)
+			}
+			data[row][col] = proof[j]
+		}
+		proofs = append(proofs, proof)
+	}
+	return
+}
+
 func getSiacoinElements(tx *txn, ids []types.SiacoinOutputID, includeProof bool) (result []explorer.SiacoinOutput, err error) {
 	var encoded []any
 	for _, id := range ids {
@@ -166,20 +211,27 @@ func getSiacoinElements(tx *txn, ids []types.SiacoinOutputID, includeProof bool)
 	}
 	defer rows.Close()
 
+	var leafIndices []uint64
 	for rows.Next() {
 		sco, err := scanSiacoinOutput(rows)
 		if err != nil {
 			return nil, fmt.Errorf("failed to scan siacoin output: %w", err)
 		}
-		if includeProof {
-			sco.StateElement.MerkleProof, err = merkleProof(tx, sco.StateElement.LeafIndex)
-			if err != nil {
-				return nil, fmt.Errorf("failed to get output merkle proof: %w", err)
-			}
-		}
 
+		leafIndices = append(leafIndices, sco.StateElement.LeafIndex)
 		result = append(result, sco)
 	}
+
+	if includeProof {
+		proofs, err := fillElementProofs(tx, leafIndices)
+		if err != nil {
+			return nil, fmt.Errorf("failed to fill siacoin output proofs: %w", err)
+		}
+		for i := range result {
+			result[i].StateElement.MerkleProof = proofs[i]
+		}
+	}
+
 	return
 }
 ```

```
**Before**
BenchmarkSiacoinOutputs/10_unspent_outputs-16         	   23709	     43972 ns/op	   21811 B/op	     324 allocs/op
BenchmarkSiacoinOutputs/100_unspent_outputs-16        	    7676	    163466 ns/op	  139756 B/op	    2127 allocs/op
BenchmarkSiacoinOutputs/1000_unspent_outputs-16       	     844	   1344622 ns/op	 1225738 B/op	   20132 allocs/op
BenchmarkSiacoinOutputs/10_siacoin_elements-16        	    9141	    129601 ns/op	   64923 B/op	    1072 allocs/op
BenchmarkSiacoinOutputs/100_siacoin_elements-16       	     753	   1571962 ns/op	  585470 B/op	    9593 allocs/op
BenchmarkSiacoinOutputs/1000_siacoin_elements-16      	      67	  15921842 ns/op	 5679791 B/op	   94665 allocs/op

**After**
BenchmarkSiacoinOutputs/10_unspent_outputs-16         	   23065	     51646 ns/op	   21810 B/op	     324 allocs/op
BenchmarkSiacoinOutputs/100_unspent_outputs-16        	    7718	    163920 ns/op	  139752 B/op	    2127 allocs/op
BenchmarkSiacoinOutputs/1000_unspent_outputs-16       	     880	   1316732 ns/op	 1225736 B/op	   20132 allocs/op
BenchmarkSiacoinOutputs/10_siacoin_elements-16        	   22579	     52932 ns/op	   40811 B/op	     486 allocs/op
BenchmarkSiacoinOutputs/100_siacoin_elements-16       	    4558	    264634 ns/op	  313874 B/op	    3052 allocs/op
BenchmarkSiacoinOutputs/1000_siacoin_elements-16      	     452	   2621011 ns/op	 2928436 B/op	   28539 allocs/op
```

It's still slower but acceptable. The initial 15ms is pretty acceptable.